random(.go): Add build tags to Go files to govet doesn't complain
The complain is about multiple main functions since every file is in the
same package main. They all have to be, otherwise `go run` doesn't work,
but they're not actually in the same program. govet thinks so, so it
complains about main being redeclared.

It's not a fatal error; the files run normally. However, it's annoying,
and I'm not sure if that being an error would suppress other stuff.

The actual solution would be to create new directories for each, but I
don't want to do that for aesthetic reasons. Instead, I can add build
tags, which makes govet realise that they're separate stuff. Since I'm
always using `go run`, the build tags don't actually change anything for
me.

It's still annoying because now I have a warning about the linter not
finding other packages for the build tag, but it's better to have a
warning rather than an error.