s2orc: Replace unique_venues.fish with Python version
The fish/fd/gzip/jq version was faster, but the Python script is easier
for others to read, and that's a real concern now.

PS: gzip CLI tool is _realy_ fast. Python's gzip is a lot slower, but so
is Go's, which surprised me. Anyway, this isn't a script that's going to
be run often, so it doesn't matter.