weekly-calendar: Show weeks between two dates
Inspired by Wait but Whys' Your Life in Weeks. Shows each week in a row,
with past days marked, then some stats at the bottom.

Meant to give me some perspective of the passage of time wrt to a
deadline.

This was written in Go because I'm using this as a greeting on Fish
shell, so Python's extra startup time was a bit annoying.

PS: This is another situation where I'm reminded why I'm not a fan of
Go: see the now/today date thing. Still, it was better to do this than
spinning up a new Rust project and remembering how that works, I
suppose.