weekly-calendar: Print todo list after the calendar
The todo list follows the usual markdown format. I don't actually do
anything to it, just the read the file and print it back. The only thing
done is remove the trailing space, and if the string is empty, I don't
print it.

This is optional, so I was forced to use the stdlib flag for it. I
couldn't be arsed to actually fetch an external library just for this.