Use RenderableType instead of ConsoleRenderable and get theme from CLI
Also defaults annotation to "Unknown" instead of omitting it. This is
because I favour pyright's strict type mode, which does this.

However, pyright also (mistakenly, IMO) infers the return type of a
function if it is not provided (given all the other types). I don't like
it, but it works, so I'm allowing the return type to be empty.