cosy: Add script to find functions without a return type annotation
Pyright allows you to omit the return annotation so it can infer the
return type. This is useful when prototyping, but I like having the
explicit return types in the end. This helps me find those cases.

Note that mypy doesn't do that, so if the project aims to be compatible
with it, it must include the explicit return types anyway.

It's a little annoying that pyright explicitly doesn't want to implement
warnings about this, claiming it's a "stylistic choice" of the project.
Anyway, this tool is simple enough since we can do everything
statically.